align ALL tdf reporting data and individual tdf reports
