+++ mofacts/client/views/experimentReporting/instructorReporting.html
@@ -21,8 +21,10 @@
       <div>
         <h3>Select Chapter</h3>
         <select id="tdf-select">
-            <option selected="selected" value="invalid" disabled="true">Select a tdf</option>
-            <option value="xml">All</option>
+            {{#if tdfs}}
+              <option selected="selected" value="invalid" disabled="true">Select a tdf</option>
+              <option value="xml">All</option>
+            {{/if}}
             {{#each tdfs}}
               <option value="{{this.fileName}}">{{this.displayName}}</option>
             {{/each}}
