+++ mofacts/client/views/versionInfo.js
@@ -1,8 +1,8 @@
 //NOTE: This file will be out of date on any machine except the one that last created the deployment tarball
 Session.set("versionInfo",
 {
-"commitID":"23046ebb91ff25f84c9634a1b8e463ce173d1935",
-"buildTimestamp":"Tue Feb 16 09:49:23 CST 2021",
+"commitID":"ff931e9671bd2d2ed2c3b16ef88c2a9c5e18fdaf",
+"buildTimestamp":"Tue Feb 16 13:09:30 CST 2021",
 "buildersUsername":"Andrew Tackett",
 "branch":"master",
 "repoURL":"git@github.com:memphis-iis/mofacts-ies.git"
